+++ core-bundle/src/Resources/contao/modules/ModuleQuicknav.php
@@ -160,7 +160,8 @@ class ModuleQuicknav extends \Module
 						'level' => ($level - 2),
 						'title' => \StringUtil::specialchars(\StringUtil::stripInsertTags($objSubpage->pageTitle ?: $objSubpage->title)),
 						'href' => $objSubpage->getFrontendUrl(),
-						'link' => \StringUtil::stripInsertTags($objSubpage->title)
+						'link' => \StringUtil::stripInsertTags($objSubpage->title),
+						'active' => ($objPage->id == $objSubpage->id || $objSubpage->type == 'forward' && $objPage->id == $objSubpage->jumpTo)
 					);
 
 					// Subpages