+++ core-bundle/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ### 4.4.0 (2017-XX-XX)
 
+ * Auto-select the active page in the quick navigation/link module (see contao/core#8587).
  * Add the contao.root_dir parameter (see #662).